separate styled components into own file, separate pages into own folder, style checkboxes
