+++ .vscode/settings.json
@@ -1,7 +1,7 @@
 {
     "workbench.colorCustomizations": {
-        "activityBar.background": "#053615",
-        "titleBar.activeBackground": "#074B1D",
-        "titleBar.activeForeground": "#EEFDF3"
+        "activityBar.background": "#352D15",
+        "titleBar.activeBackground": "#4A3F1E",
+        "titleBar.activeForeground": "#FCFBF8"
     }
 }
