+++ code/src/components/Header.js
@@ -1,99 +1,18 @@
 import React from "react"
 import { useSelector } from "react-redux"
 import Moment from "react-moment"
-import styled from "styled-components"
 
-const StyledHeader = styled.header`
-    margin: 1.5rem;
-    padding: 1.3rem;
-    background: linear-gradient(200deg, #e3c5f4 0%, #c56bf9 55%);
-    border-radius: 10px;
-    color: white;
-    font-family: "Poppins", sans-serif;
-`
+import {
+    StyledHeader,
+    HeaderDetails,
+    HeaderTitle,
+    CurrentTasks,
+    HeaderGreeting,
+    HeaderDate,
+    ButtonContainer,
+    TaskFilterButton } from "styles/StylesForComponents"
 
-const HeaderDetails = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-`
-
-const HeaderTitle = styled.h1`
-    font-size: 1.7rem;
-    margin-bottom: 0.5rem;
-    text-transform: uppercase;
-    font-weight: bold;
-    width: 30%;
-
-    @media screen and (min-width: 1024px) {
-        font-size: 2.2rem;
-    }
-`
-
-const CurrentTasks = styled.p`
-    font-size: 0.8rem;
-    width: 30%;
-    background: white;
-    padding: 6px;
-    border-radius: 10px;
-    color: #060624;
-    text-align: center;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    box-shadow: 0px 3px 6px 0px #00000081;
-
-    @media screen and (min-width: 1024px) {
-        font-size: 0.9rem;
-    }
-`
-
-const HeaderGreeting = styled.h2`
-    font-size: 1.1rem;
-    font-weight: normal;
-    margin-bottom: 3px;
-
-    @media screen and (min-width: 1024px) {
-        font-size: 1.4rem;
-    }
-`
-
-const HeaderDate = styled.p`
-    font-size: 0.9rem;
-
-    @media screen and (min-width: 1024px) {
-        font-size: 1rem;
-    }
-`
-
-const ButtonContainer = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    margin-top: 2rem;
-`
-
-const TaskFilterButton = styled.button`
-    appearance: none;
-    border: none;
-    padding: 5px;
-    border-radius: 10px;
-    font-size: 0.9rem;
-    margin: 0;
-    background: transparent;
-    color: white;
-    font-family: "Poppins", sans-serif;
-
-    &:hover {
-        text-decoration: underline;
-        cursor: pointer;
-    }
-
-    @media screen and (min-width: 1024px) {
-        font-size: 1.1rem;
-    }
-`
-
+    
 const Header = (props) => {
     const {
         allTasks,
