+++ code/src/components/TaskItemPriority.js
@@ -1,58 +1,11 @@
 import React from "react"
-import styled from "styled-components"
 
+import {
+    PriorityItem,
+    PriorityTitle,
+    PriorityButtonContainer,
+    PriorityButton } from "styles/StylesForComponents"
 
-const PriorityItem = styled.div`
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    margin: 1rem 0 2rem;
-
-    @media screen and (min-width: 1024px) {
-        margin: 2rem 0 3rem;
-    }
-`
-
-const PriorityTitle = styled.p`
-    font-size: 0.9rem;
-
-    @media screen and (min-width: 1024px) {
-        font-size: 1.1rem;
-        margin-bottom: 5px;
-    }
-`
-
-const PriorityButtonContainer = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-`
-
-const PriorityButton = styled.button`
-    appearance: none;
-    border: 2px solid;
-    border-color: transparent;
-    margin: 10px 10px 0 0;
-    padding: 5px;
-    border-radius: 10px;
-    font-size: 0.8rem;
-    width: 60px;
-    color: #000000ec;
-
-    &:hover,
-    &:active,
-    &:focus {
-        outline: none;
-        border-color: black;
-        opacity: 0.7;
-    }
-
-    @media screen and (min-width: 1024px) {
-        margin: 10px 15px 0 0;
-        width: 70px;
-        font-size: 1rem;
-    }
-`
 
 const TaskItemPriority = ({ priority, setPriority }) => {
     const priorities = [ "Urgent", "Normal", "Low"]
