+++ code/src/index.css
@@ -33,7 +33,6 @@ a:focus {
   color: white;
 }
 
-
 .urgent-tag {
   background: linear-gradient(190deg, rgba(238,215,215,1) 0%, rgba(255,105,97,1) 100%);
 }
@@ -44,4 +43,13 @@ a:focus {
 
 .low-tag {
   background: linear-gradient(194deg, rgba(237,244,235,1) 24%, rgba(190,229,176,1) 100%);
+}
+
+input {
+  accent-color: #b66fdf;
+}
+
+input[type="checkbox"] {
+  width: 1.3rem;
+  height: 1.3rem;
 }
