+++ code/src/pages/StartPage.js
@@ -1,74 +1,16 @@
 import React, { useState } from "react"
 import { useDispatch } from "react-redux"
 import { useNavigate } from "react-router-dom"
-import styled from "styled-components"
-
-import StyledButton from "./StyledButton"
 
 import taskitems from "reducers/taskitems"
+import StyledButton from "../styles/StyledButton"
+import { 
+    StyledStartPage,
+    StartMainTitle,
+    StartSubTitle,
+    StyledInput,
+    StartInput } from "styles/StylesForPages"
 
-const StyledStartPage = styled.section`
-    width: 100vw;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    text-align: center;
-    padding: 1rem;
-    font-family: "Inter", sans-serif;
-`
-
-const StartMainTitle = styled.h1`
-    margin-bottom: 1rem;
-    font-size: 3.5rem;
-    text-transform: uppercase;
-
-    @media screen and (min-width: 1024px) {
-        font-size: 4rem;
-        margin-bottom: 1.5rem;
-    }
-`
-
-const StartSubTitle = styled.h2`
-    margin-bottom: 2rem;
-
-    @media screen and (min-width: 1024px) {
-        font-size: 2rem;
-    }
-`
-
-const StyledInput = styled.label`
-    display: flex;
-    flex-direction: column;
-    font-size: 1.1rem;
-
-    @media screen and (min-width: 1024px) {
-        font-size: 1.5rem;
-    }    
-`
-
-const StartInput = styled.input`
-    appearance: none;
-    border: transparent;
-    background: whitesmoke;
-    margin-top: 7px;
-    border-radius: 10px;
-    padding: 5px;
-    margin-bottom: 2rem;
-    text-align: center;
-    font-size: 1rem;
-
-    &:hover,
-    &:focus {
-        outline: none;
-        border: 1px solid purple;
-    }
-
-    @media screen and (min-width: 1024px) {
-        font-size: 1.3rem;
-    }  
-`
 
 const StartPage = () => {
     const [nameInputValue, setNameInputValue] = useState("")
