+++ code/src/styles/StyledButton.js
@@ -1,4 +1,4 @@
-import styled from "styled-components"
+import styled from "styled-components/macro"
 
 const StyledButton = styled.button`
     appearance: none;
